Fix bug causing deadlock between FreeRTOS and XC
Discovered bug which caused deadlock in demo app with 2 or fewer
FreeRTOS cores.  Bug was determined to be caused by an XC and FreeRTOS
deadlock.  Bug was fixed by changing which tasks are combinable.

The bug is caused by having a single select block with at least 2
devices.  One device, device A,  must have a case waiting on
soc_peripheral_function_code_rx and the other, device B, a call to
soc_peripheral_tx_dma_xfer within one of the cases.
In FreeRTOS, the core that the ISR is registered to for the dma transfer
from device B, must be running a task that calls
soc_peripheral_function_code_tx to device A, which masks interrupts for
a channel transfer.  If a case within the select block is executing,
then the soc_peripheral_function_code_rx will not be immediate recieved,
which is acceptable; however, if the device B case is running and calls
soc_peripheral_tx_dma during this time, then we are in deadlock.  Device
A cannot respond to the function code because we are not at a waiteu
instruction in XC.  The call to soc_peripheral_tx_dma will wait until
the interrupt is handled on the FreeRTOS side.  Meanwhile, the FreeRTOS
side is waiting, with interrupts masked, for the XC device to respond to
the command.