Hack to get WiFi working
Having trouble reading firmware out of flash.
Firmware needed to be moved into the function that reads it out of the
array due to memory errors.

SPI driver timing adjusted without a good understanding of what this
really did, but it seems to have fixed the issue.

Root causes for the failrues need to be investigated and real fixes need
to be put in place and these hacks removed.