Updating FAT filesystem code - WIP
Updated the FreeRTOS Plus FAT driver to use the new QSPI flash device as
well as to use 4096 byte sectors. This almost works, but there are many
places in the FreeRTOS Plus FAT library that use the hardcoded sector
size of 512 bytes. Thinking about abandoning FreeRTOS Plus FAT.

Updated the filesystem example app to use the new FatFs software service
module. This is working to boot the WFM200 module and uses much less RAM
(code, data, and heap) than FreeRTOS plus FAT. Unfortunately it is not
directly compatible with FreeRTOS Plus FAT so the FTP server for now is
disabled.

The filesystem application is currently a mess. FreeRTOS Plus FAT needs
to be removed completely and a new FTP server added to work with FatFs.