Fix MQTT client cleanup behavior with TLS
The TLS MQTT use case assumes that the socket and TLS session are
handled by a task that spawns the MQTT client.  The MQTT client should
not free the socket for the TLS case.  When the MQTT Network has an
ssl_ctx, it a call to FreeRTOSDisconnect will no longer close the
socket.

If both the MQTT client and socket handling task call
FreeRTOS_closesocket, an assertion will occur in vPortFree, where the
second call will try to free a block that is not allocated.