Improve the WiFi connection manager
Added an application callback to the WiFi connection manager. This
allows the application to move the connection manager between station
and soft AP modes. Also added two new functions to allow the application
to disconnect from an AP and stop the soft AP.

Minor updates to the IoT WiFi functions, mostly to the soft AP functions
so that they match recent updates to the connect/disconnect functions.

In the explorer example app:
 - Fixed the thruput_test so that it completes.
 - Suppressed a lot of debug printing.
 - Using the new WiFi connection manager application callback. Added a
   test mode to alternate between the two modes at a regular interval.