Update WiFi to load PDS config from filesystem
The AFR IoT WiFi function WIFI_On() will now check to see if the WF200
PDS config data is in the filesystem, and will use it if it is.
Otherwise it will fallback on the compiled in brd8023a config for the
WFM200 module.

WiFi connection manager also updated to not use the DHCP server if it
was not compiled in.