Change default WF200 FW location from flash to RAM
The default WF200 firmware readback function now loads the firmware from
the array located in the WF200 driver, rather than from the start of the
data partition in flash. Since there are other things that store data in
flash, it has become more useful to store the firmware in a FAT
filesystem in the flash when the firmware needs to be located in flash;
thus the previous default was no longer useful. Because the focus moving
forward is on xcore.ai rather than xcore200, reading the firmware directly
out of RAM is now actually possible, therefore it is now the new default.