Add a USB example app
This example app is based on the HID example from lib_xud. Only portions
of it use the RTOS USB driver. Other parts call lib_xud functions
directly. This is only being used to test the new RTOS driver, and is
not currently a good example. It also does not work correctly with
Linux.