Attempt to improve receipt of USB setup packets
lib_xud should now be prepared to receive setup packets immediately
following transfer of ZLP status from the device to the host. There is
still a time gap between receipt of a ZLP status from the host preparing
for the next setup packet from the host. If a setup packet comes in
during this gap, this could be a problem. This is under investigation.