Make USB HID example works on XCORE200
The USB HID example now works on the XCORE200 Mic Array board.
It works both with the DEV/XS3 and master branches of lib_xud.
The master branch needed a slight modification to the function that sets
up buffer transfers to the host.

USB tile defines added to the board support cmake files.