Correctly check the TUSB speed config option
The XCORE device port now correctly checks the CFG_TUSB_RHPORT0_MODE
setting for the speed option to determine whether to run lib_xud in full
or high speed mode. Previously it checked BOARD_DEVICE_RHPORT_SPEED
which is not technically part of the library.

The audio test demo is also fixed to correctly set the speed. When the
Tiny USB stack knows that the device is running in high speed mode, it
then only needs data from the application once every 1 ms, rather than
once every 125 us. This allows the application code to be the same for
both high and full speed modes.