multitile_build.mk moved into common cmake_utils
This moves all the common makefile noise out of each application's
makefile and into a common shared makefile include file. It is used for
building applications with independently built tiles. If an application
does not need that, then it does not need to include this file.
Applications may continue to choose to do this manually as well.