Add new thread to RTOS I2C driver for callbacks
A second thread has been added to the RTOS I2C driver that is
responsible for calling the application's provided callback functions.
This ensures that the "HIL" thread never blocks (from the RTOS's point
of view). This ensures that the I2C HIL thread will not resume on a core
that has any interrupts enabled on it.