Update USB driver to resolve interrupt issues
The USB driver and stack should now be initialized prior to starting the
RTOS scheduler. This ensures that the lib_xud main task is placed on a
core immediately when the scheduler starts, ensuring that no other tasks
have enabled any interrupts on it first, which has been causing
problems.