Fix problem with stalling USB EP0
This fix relies on a change to lib_xud that has not yet been merged into
the official lib_xud repo. In this interim, the xud submodule is
pointing to a fork with the change.

Device control over USB example has been updated to stall EP0 when a
command for an unregisted resource comes in earlier than it had been.
This is testing the USB stall fix, and causes issues without it.