Add new ISR callbacks to the RTOS SWMem driver
This allows applications to provide software memory handlers that are
called from within the software memory interrupt service routines,
rather than from within a thread. This will speed up accesses, provided
the handler does not need to be run from within a thread.

It also allows for the application to provide both ISR and thread
callbacks. If the ISR callbacks are unable to satisfy a request, then
the driver falls back on the thread callbacks.