Fix busy loops in example app pipeline stages
The busy loops in the independent tile example app's pipeline stages
assumed the number of cycles per iteration to be 3. This is true with
xcc, but for whatever reason it is 4 with xClang. The loops have been
rewritten to not rely on this, and are now timed with the hardware
reference time.