Add filter callbacks to RTOS I2S driver
This gives the application a way to filter the i2s audio between the
buffer and the interface from within the I2S i/o thread. If these
callbacks are not specified then the behavior is the same as it has
been. New functions have been added to set these filter callbacks, so
the API is backwards compatible.

Currently only the send filter is implemented.